Move type name shortening to bevy_utils
